requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-generic-json-views==0.8
django-jsonview==1.1.0
django-modeltranslation==0.12.1
django-rosetta==0.8.1
django-solo==1.1.3
idna==2.6
microsofttranslator==0.8
Pillow==5.0.0
pluggy==0.6.0
polib==1.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
pytz==2018.3
requests==2.18.4
six==1.10.0
tox==2.9.1
urllib3==1.22
virtualenv==15.1.0
